+++ DesignerMacros/src/rdbms/key-automation-extensions/attribute-on-loaded/attribute-on-loaded.ts
@@ -2,29 +2,28 @@
 /// <reference path="../_common/constants.ts" />
 
 function execute(): void {
-    if (element.getPackage().specialization !== "Domain Package" ||
+    if (!element.hasMetadata(metadataKey.association) ||
+        element.getPackage().specialization !== "Domain Package" ||
         !element.getPackage().hasStereotype(relationalDatabaseId)
     ) {
         return;
     }
 
-    const associationTarget = element.getStereotype(foreignKeyStereotypeId)?.getProperty(foreignKeyStereotypeAssociationProperty)?.getValue() as MacroApi.Context.IAssociationApi;
-    associationTarget.isSourceEnd();
-    associationTarget.getOtherEnd().id
-
-    const existingAssociation = element.getMetadata(metadataKey.association);
-
-    if (associationTarget == null && existingAssociation != null) {
+    const associationTarget = element.getStereotype(foreignKeyStereotypeId)?.getProperty(foreignKeyStereotypeAssociationProperty)?.getValue();
+    if (associationTarget == null) {
         if (!element.hasStereotype(foreignKeyStereotypeId)) {
             element.addStereotype(foreignKeyStereotypeId);
         }
 
-        const stereotype = element.getStereotype(foreignKeyStereotypeId);
-        stereotype.getProperty(foreignKeyStereotypeAssociationProperty).setValue(existingAssociation);
+        const associationId = element.getMetadata(metadataKey.association);
+        element
+            .getStereotype(foreignKeyStereotypeId)
+            .getProperty(foreignKeyStereotypeAssociationProperty)
+            .setValue(associationId);
     }
 
-    if (existingAssociation && !element.getMetadata("fk-original-name")) {
-        element.setMetadata("fk-original-name", element.getName());
+    if (!element.hasMetadata(metadataKey.fkOriginalName)) {
+        element.setMetadata(metadataKey.fkOriginalName, element.getName());
     }
 }